# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Pesa AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Business {
  id: string;
  name: string;
  personaName: string;
  category: string;
  phone: string;
  paybillNumber?: string | null;
  whatsappPhoneNumberId?: string | null;
  createdAt: string;
}

export interface Account {
  id: string;
  email: string;
}

export interface Subscription {
  id: string;
  businessId: string;
  plan: string;
  status: string;
  priceKES: number;
  trialEndsAt?: string | null;
  currentPeriodEnd: string;
  createdAt: string;
}

export interface Product {
  id: string;
  businessId: string;
  name: string;
  description?: string | null;
  price: number;
  stockQty: number;
  imageUrl?: string | null;
  active: boolean;
  createdAt: string;
}

export interface OrderItem {
  productId: string;
  productName: string;
  qty: number;
  unitPrice: number;
}

export interface Order {
  id: string;
  businessId: string;
  customerId: string;
  customerPhone: string;
  customerName?: string | null;
  items: OrderItem[];
  totalKES: number;
  status: string;
  deliveryAddress?: string | null;
  createdAt: string;
}

export interface SalesTrendPoint {
  date: string;
  revenue: number;
  orderCount: number;
}

export interface TopProduct {
  productId: string;
  productName: string;
  revenue: number;
  qty: number;
}

export interface SalesSummary {
  totalRevenue: number;
  totalOrders: number;
  avgOrderValue: number;
  trend?: SalesTrendPoint[];
  topProducts?: TopProduct[];
  advancedAnalyticsLocked?: boolean | null;
}

export interface MpesaStatus {
  connected: boolean;
  shortcodeMasked?: string | null;
}

export interface ChatMessage {
  id: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatReply {
  replyText: string;
}

export interface Report {
  id: string;
  businessId: string;
  reason: string;
  details?: string | null;
  reporterContact?: string | null;
  status: string;
  createdAt: string;
}

export interface ReportGroup {
  business?: Business;
  openCount: number;
  reports: Report[];
}

export interface BusinessWithSubscription {
  id: string;
  name: string;
  personaName: string;
  phone: string;
  category: string;
  paybillNumber?: string | null;
  whatsappPhoneNumberId?: string | null;
  createdAt: string;
  subscription: Subscription;
}

export interface AdminStats {
  totalBusinesses: number;
  activeBusinesses: number;
  trialingBusinesses: number;
  suspendedBusinesses: number;
  totalRevenue: number;
  revenueThisMonth: number;
  openReports: number;
  whatsappConnected: number;
}

export interface MeResult {
  authenticated: boolean;
  isAdmin?: boolean | null;
  account?: Account;
  business?: Business;
  subscription?: Subscription;
}

export interface AuthResult {
  business: Business;
  account: Account;
  subscription: Subscription;
}

export interface SignupInput {
  businessName: string;
  category: string;
  businessPhone: string;
  paybillNumber?: string;
  email: string;
  password: string;
  consent: boolean;
  plan?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface BusinessUpdate {
  name?: string;
  category?: string;
  paybillNumber?: string;
  whatsappPhoneNumberId?: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  stockQty?: number;
  imageUrl?: string;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  stockQty?: number;
  imageUrl?: string;
  active?: boolean;
}

export interface ProductImportInput {
  rows: ProductInput[];
}

export interface OrderStatusUpdate {
  status: string;
}

export interface MpesaPayInput {
  phone: string;
}

export interface MpesaConnectInput {
  consumerKey: string;
  consumerSecret: string;
  passkey: string;
  shortcode: string;
}

export interface PlanInput {
  plan: string;
}

export interface SubscriptionChargeInput {
  phone?: string;
}

export interface ChatMessageInput {
  customerPhone: string;
  customerName?: string;
  message: string;
}

export interface ReportInput {
  businessId: string;
  reason: string;
  details?: string;
  reporterContact?: string;
}

export interface ReportStatusUpdate {
  status: string;
}

export interface AdminChargeInput {
  phone?: string;
}