# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pesa AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports:
    post:
      operationId: createReport
      tags: [reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: OK
...
  /admin/reports/{id}:
    patch:
      operationId: adminUpdateReport
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportStatusUpdate"
      responses:
        "200":
          description: OK
components:
  schemas:
...
    ReportInput:
      type: object
      properties:
        businessId:
          type: string
        reason:
          type: string
        details:
          type: string
        reporterContact:
          type: string
      required:
        - businessId
        - reason
    ReportStatusUpdate:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...